requirements.txt
streamlit==1.38.0
llama-index==0.10.68
llama-index-embeddings-openai==0.1.11
llama-index-llms-openai==0.1.29
chromadb==0.5.3
llama-index-vector-stores-chroma==0.1.10
pypdf==4.3.1
python-dotenv==1.0.1
openai==1.40.0